Move freemarker class to freemarker package
